+++ NEWS.md
@@ -54,11 +54,14 @@
 * Added routine for histogram smoothing with LPS.
 * Added routines to estimate the incubation period distribution.
 
-### Version 1.2.5 ### (**2024-01-08**)
+### Version 1.3.0 ### (**2024-03-04**)
 
 * Version name: "Extension to nowcasting".
 * Added nowcasting routine and associated S3 method.
+* Added nowcastingR routine and associated S3 method.
 * Added incidence and mortality datasets for Belgium.
+* Corrected Rsd formula in KerRpostmap.cpp file.
+* Added 0 as a lower bound on ylim for S3_Rt_plot.R.